refactor: web3 state
