+++ packages/plugins/Flow/src/storage/index.ts
@@ -4,7 +4,9 @@ import type { ScopedStorage } from '@masknet/shared-base'
 export const StorageDefaultValue = {
     user: null as User | null,
 }
+
 export let storage: ScopedStorage<typeof StorageDefaultValue> = null!
+
 export function setupStorage(_: typeof storage) {
     storage = _
 }
